Update of the homogeneous solution loop
Now the data for the Dirac's delta response is a mutable struct and it is updated for each DOF in the loop for DOFs. Then, all the solution due to Dirac's delta (for impulses in all DOFs) is evaluated alongside the homogeneous response in a single loop. Hence, modifications were done in all four main files - solver.jl, process.jl, structs.jl and initial_conditions.jl (the n_excitedDOF variable was taken out, since it was not being used).