Update of the Dirac's delta solution
Since the solution due to all Dirac's deltas and the homogeneous solution must be evaluated in a single loop along the points of time, the function that evaluates the response due to Dirac's delta was updated to receive the initial conditions constant vectors C1 and C2, furthermore, the vectors v_j are calculated and stored in a matrix. The late function for the response due to Dirac's delta was split in two - one for the case of complex-conjugacy between F11 and CbF and another function for the opposite case.